# src/config/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		},
		"schemas": {
			"NoteCategory": {},
			"CategoryType": {
				"$ref": "#/components/schemas/NoteCategory"
			},
			"NoteResponse": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string"
					},
					"content": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					},
					"category": {
						"$ref": "#/components/schemas/CategoryType",
						"nullable": true
					},
					"version": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"content",
					"userId",
					"version",
					"createdAt",
					"updatedAt"
				]
			},
...
			"NoteListResponse": {
				"type": "object",
				"properties": {
					"notes": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/NoteResponse"
						}
					}
				},
				"required": [
					"notes"
				]
			},
			"CreateNoteResponse": {
				"type": "object",
				"properties": {
					"message": {
						"type": "string"
					},
					"note": {
						"$ref": "#/components/schemas/NoteResponse"
					}
				},
				"required": [
					"message",
					"note"
				]
			},
			"CreateNoteDTO": {
				"type": "object",
				"properties": {
					"content": {
						"type": "string"
					},
					"category": {
						"$ref": "#/components/schemas/CategoryType"
					}
				},
				"required": [
					"content"
				]
			},
...
		}
	},
	"info": {
		"title": "Notes API",
		"version": "1.0.0",
		"description": "A RESTful API for note-taking with user authentication, versioning, and full-text search"
	},
	"paths": {
		"/notes/{noteId}/versions": {
			"get": {
				"operationId": "GetNoteVersions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"versions": {
											"items": {},
											"type": "array"
										}
									},
									"required": [
										"versions"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Retrieves all versions of a note, ordered by version number (newest first)",
				"tags": [
					"Versions"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "noteId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/notes/{noteId}/revert/{versionNumber}": {
			"post": {
				"operationId": "RevertToVersion",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"revertedFromVersion": {
											"type": "number",
											"format": "double"
										},
										"note": {
											"$ref": "#/components/schemas/NoteResponse"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"revertedFromVersion",
										"note",
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Reverts a note to a specific version number",
				"tags": [
					"Versions"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "noteId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "versionNumber",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/notes": {
			"post": {
				"operationId": "CreateNote",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CreateNoteResponse"
								}
							}
						}
					}
				},
				"description": "Creates a new note for the authenticated user",
				"tags": [
					"Notes"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateNoteDTO"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "ListNotes",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NoteListResponse"
								}
							}
						}
					}
				},
				"description": "Lists all notes for the authenticated user (owned + shared)",
				"tags": [
					"Notes"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:8080/api",
			"description": "Development server"
		}
	]
}